backend/requirements.txt
Flask
Flask-CORS
opencv-python
scikit-image
Pillow
